refactor: put gsap rotate logic in separate file
